Adding makefile
Since the project is required to be worked under the Apache© Ant the
build.xml is fulfilled today. It creates the ./bin folder as the build
folder. Then all compiled classes are redirected to that folder.
The target "run" still doesn't work so far.
The interface class(the impl package) should be compiled into jar in the
future.

Signed-off-by: Chen, Yiqun