The implementation of server methods
Taking assumption that remote call should be passed by reference the
client interfaces are also implemented on the server side.
Some further comments regarding methods to generate the document should
be added.
The asynchronized timer ought to be added on the server side to provide
constant refresh of randomly generated position of fly.

Signed-off-by: Chen, Yiqun