[Reconstruction]The reconstruction of the project
As I've guessed the IGameClient should be implemented by the client
side. Hopefully this time I'm making it right.
According to the specification, I conclude the following requirements:
---- Server Side ----
1. GameServer.login and GameServer.logout have been fully implemented
which ought to be invoked by the client.
2. GameServer.huntFly is set to be invoked in the callback function from
client upon the event that the picture of the fly is clicked.

---- Client Side ----
1. GameClient.receiveFlyPosition is invoked in the asynchronized timer
callback method. The server should pass two new randomly generated
number to the client.
2. GameClient.receiveFlyHunted is invoked by the server once it has
decided to confer the bonus to the given player. The player name and new
score should be passed to the client to refresh its GUI.

Side notes: Once the fly was caught the server timer should be updated
so that it won't generate new random number in a shorter time interval.