Apply reliable synchronization
The modified codes apply the try lock mechanism to ensure the lock won't
be hold for long. Once the attempt to acquire lock fails the tryLock()
returns immediately. So we can force the players who failed to access
the lock to quit ASAP and left all refresh works to the thread which
successfully holds the lock.

Signed-off-by: Chen, Yiqun